Number of deaths and population by sex and age groups for Canada in 2023

Number of deaths and population by sex and age groups for Canada in 2023

Usage

CanadaDeaths2023

Format

`CanadaDeaths2023` A data frame with 21 rows and 6 columns:

start

age in years at start of age interval

end

age in years at end of age interval

population.female

female population in the age interval on 2023-07-01

population.male

male population in the age interval on 2023-07-01

deaths.female

number of females in the age interval dying during 2023

deaths.male

number of males in the age interval dying during 2023

Source

Statistics Canada; deaths https://doi.org/10.25318/1310070901-eng; population https://doi.org/10.25318/1710000501-eng; downloaded 2025-01-22.

Examples

CanadaDeaths2023
#>    start end population.female population.male deaths.female deaths.male
#> 1      0   4            914872          960210           849        1105
#> 2      5   9           1034125         1088689            98          98
#> 3     10  14           1063378         1125824           121         145
#> 4     15  19           1080547         1150565           284         523
#> 5     20  24           1241956         1354011           523         991
#> 6     25  29           1391946         1517489           782        1603
#> 7     30  34           1456831         1532733          1023        2109
#> 8     35  39           1387063         1425708          1241        2495
#> 9     40  44           1331021         1346434          1441        2787
#> 10    45  49           1234659         1224346          1917        3224
#> 11    50  54           1229275         1202081          2810        4772
#> 12    55  59           1288417         1261813          4588        7687
#> 13    60  64           1358461         1323303          7591       11936
#> 14    65  69           1225282         1154724         10760       15820
#> 15    70  74            998233          912977         13863       19227
#> 16    75  79            772554          685170         18353       23007
#> 17    80  84            507257          409343         21182       23912
#> 18    85  89            319764          218994         24885       23276
#> 19    90  94            174701           90629         25518       17556
#> 20    95  99             56801           19805         14659        6271
#> 21   100 Inf              9364            2129          4163        1020