Sex-specific mortality rates for Canada in 2023 by age groups

Sex-specific mortality rates for Canada in 2023 by age groups

Usage

CanadaMortality

Format

`CanadaMortality` A data frame with 20 rows and 4 columns:

start

age in years at start of age interval

end

age in years at end of age interval

q.female

female mortality rate, annual probability of dying in the age interval

q.male

male mortality rate, annual probability of dying in the age interval

Source

Statistics Canada; https://doi.org/10.25318/1310071001-eng; downloaded 2025-01-22.

Examples

CanadaMortality
#>    start end q.female q.male
#> 1      0   0   0.0042 0.0051
#> 2      1   4   0.0002 0.0002
#> 3      5   9   0.0001 0.0001
#> 4     10  14   0.0001 0.0001
#> 5     15  19   0.0003 0.0005
#> 6     20  24   0.0004 0.0007
#> 7     25  29   0.0006 0.0011
#> 8     30  34   0.0007 0.0014
#> 9     35  39   0.0009 0.0017
#> 10    40  44   0.0011 0.0021
#> 11    45  49   0.0016 0.0026
#> 12    50  54   0.0023 0.0040
#> 13    55  59   0.0036 0.0061
#> 14    60  64   0.0056 0.0090
#> 15    65  69   0.0088 0.0137
#> 16    70  74   0.0139 0.0210
#> 17    75  79   0.0238 0.0335
#> 18    80  84   0.0418 0.0584
#> 19    85  89   0.0779 0.1062
#> 20    90 Inf   0.1840 0.2208