Sex-specific mortality rates for Canada in 2022–2023 by single years of age

Sex-specific mortality rates for Canada in 2022–2023 by single years of age

Usage

CanadaMortalitySingleYear

Format

`CanadaMortalitySingleYear` A data frame with 110 rows and 3 columns:

age

age in years

q.female

female mortality rate, probability of dying during the year

q.male

male mortality rate, probability of dying during the year

Source

Statistics Canada; https://doi.org/10.25318/1310011401-eng; downloaded 2025-01-23.

Examples

head(CanadaMortalitySingleYear)
#>   age q.female  q.male
#> 1   0  0.00406 0.00477
#> 2   1  0.00023 0.00030
#> 3   2  0.00017 0.00022
#> 4   3  0.00014 0.00017
#> 5   4  0.00011 0.00013
#> 6   5  0.00010 0.00011
tail(CanadaMortalitySingleYear)
#>     age q.female  q.male
#> 105 104  0.40262 0.44999
#> 106 105  0.42511 0.46987
#> 107 106  0.44672 0.48866
#> 108 107  0.46732 0.50626
#> 109 108  0.48676 0.52263
#> 110 109  0.50496 0.53776