Calculate age-specific mortality rates

Calculate age-specific mortality rates from number of deaths and population by single years of age or age groups.

Usage

mortalityRates(
  population,
  deaths,
  start = 0:(n - 1),
  end = start,
  day = 182,
  leapyear = FALSE,
  a0 = 0.7,
  a1 = 0.6
)

Arguments

population

number of individuals in the age interval

deaths

number of deaths in the age interval during the calendar year

start

the age in years at the start of each age interval; if missing, then the default is single years from 0 to n - 1, where n is the number of intervals

end

the age in years at the end of each interval; the default is start, which produces single-year intervals; if the last interval is open-ended then end[n] can be Inf

day

day of the year on which population was measured; may be a number, defaulting to 182 (July 1 in a non-leap-year), or any value that can be interpreted by the yday() function in the lubridate package

leapyear

TRUE if day occurs in a leap-year (default is FALSE)

a0

proportion of deaths that take place during the first half of the first year of life, default 0.7; used only if the first age interval is the single year age 0

a1

proportion of deaths that take place during the first half of the second year of life, default 0.6; used only if the second age interval is the single year from age 1

Value

a data frame with, for single-year data, columns for age and the one-year mortality.rate, or, for age-grouped data, columns for the start and end in years of each age group, and the one-year mortality.rate.

Details

Calculates the single-year mortality rate at each age or in each age group. First, population size on January 1 is approximated by adding the interpolated number of deaths between January 1 and the date at which population is measured, assuming that deaths are evenly spread throughout the year (with the exception of the first and second single years of life, where a0 and a1 proportion of deaths, respectively, are assumed to occur during the first half of each of these years of life).

See also

Author

John Fox jfox@mcmaster.ca

Examples

(mortality2023 <- with(CanadaDeaths2023,  
                       mortalityRates(population.female, deaths.female,
                             start=start, end=end, day="2023-07-02")))
#>    start end mortality.rate
#> 1      0   4   0.0009275671
#> 2      5   9   0.0000947616
#> 3     10  14   0.0001137818
#> 4     15  19   0.0002627952
#> 5     20  24   0.0004210210
#> 6     25  29   0.0005616452
#> 7     30  34   0.0007019620
#> 8     35  39   0.0008942950
#> 9     40  44   0.0010820402
#> 10    45  49   0.0015514477
#> 11    50  54   0.0022832834
#> 12    55  59   0.0035546127
#> 13    60  64   0.0055723295
#> 14    65  69   0.0087431570
#> 15    70  74   0.0137915118
#> 16    75  79   0.0234766454
#> 17    80  84   0.0409016028
#> 18    85  89   0.0749005414
#> 19    90  94   0.1360996611
#> 20    95  99   0.2285092327
#> 21   100 Inf   0.3635425985
with(mortality2023, lifetable(mortality2023$mortality.rate,
                              start=start, end=end))
#>                 l[x]  d[x]    p[x]    q[x]   L[x]    T[x]      e[x]
#> 0 - 4 years   100000   463 0.99537 0.00463 498840 8413655 84.136550
#> 5 - 9 years    99537    47 0.99953 0.00047 497570 7914815 79.516311
#> 10 - 14 years  99490    57 0.99943 0.00057 497310 7417245 74.552669
#> 15 - 19 years  99433   130 0.99869 0.00131 496840 6919935 69.593948
#> 20 - 24 years  99303   209 0.99790 0.00210 495995 6423095 64.681782
#> 25 - 29 years  99094   278 0.99719 0.00281 494775 5927100 59.812905
#> 30 - 34 years  98816   346 0.99650 0.00350 493215 5432325 54.974144
#> 35 - 39 years  98470   439 0.99554 0.00446 491250 4939110 50.158525
#> 40 - 44 years  98031   529 0.99460 0.00540 488830 4447860 45.371974
#> 45 - 49 years  97502   754 0.99227 0.00773 485625 3959030 40.604603
#> 50 - 54 years  96748  1099 0.98864 0.01136 480990 3473405 35.901569
#> 55 - 59 years  95649  1688 0.98235 0.01765 474025 2992415 31.285377
#> 60 - 64 years  93961  2589 0.97245 0.02755 463335 2518390 26.802503
#> 65 - 69 years  91372  3925 0.95704 0.04296 447045 2055055 22.491080
#> 70 - 74 years  87447  5866 0.93292 0.06708 422570 1608010 18.388395
#> 75 - 79 years  81581  9137 0.88800 0.11200 385060 1185440 14.530834
#> 80 - 84 years  72444 13652 0.81155 0.18845 328090  800380 11.048258
#> 85 - 89 years  58792 18957 0.67755 0.32245 246565  472290  8.033236
#> 90 - 94 years  39835 20667 0.48119 0.51881 147510  225725  5.666499
#> 95 - 99 years  19168 13929 0.27331 0.72669  61015   78215  4.080499
#> 100+ years      5239  4692 0.10444 0.89556  14465   17200  3.283069